Extract generate_list_and_response_models to module
