+++ crudcrate-derive/src/codegen/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod api_struct;
 pub mod create;
 pub mod list;
+pub mod list_response;
 pub mod merge;
 pub mod response;
 pub mod shared;
