+++ crudcrate-derive/src/lib.rs
@@ -14,7 +14,7 @@ mod traits;
 use proc_macro::TokenStream;
 use quote::{format_ident, quote};
 use syn::{DeriveInput, parse_macro_input};
-use traits::crudresource::structs::{CRUDResourceMeta, EntityFieldAnalysis};
+use traits::crudresource::structs::CRUDResourceMeta;
 
 fn extract_active_model_type(input: &DeriveInput, name: &syn::Ident) -> proc_macro2::TokenStream {
     let mut active_model_override = None;
@@ -161,76 +161,6 @@ pub fn to_list_model(input: TokenStream) -> TokenStream {
 ///
 /// Key attributes: `api_struct`, `generate_router`, `exclude()`, `join()`, `on_create/update`.
 /// See crate documentation for full attribute reference and examples.
-/// Generate list and response models
-fn generate_list_and_response_models(
-    input: &DeriveInput,
-    api_struct_name: &syn::Ident,
-    struct_name: &syn::Ident,
-    field_analysis: &EntityFieldAnalysis,
-) -> (proc_macro2::TokenStream, proc_macro2::TokenStream) {
-    // Generate List model
-    let list_name = format_ident!("{}List", api_struct_name);
-    let raw_fields = fields::extract_named_fields(input);
-    let list_struct_fields = crate::codegen::models::list::generate_list_struct_fields(&raw_fields);
-    let list_from_assignments =
-        crate::codegen::models::list::generate_list_from_assignments(&raw_fields);
-    let list_from_model_assignments =
-        crate::codegen::models::list::generate_list_from_model_assignments(field_analysis);
-
-    let list_derives =
-        quote! { Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize, utoipa::ToSchema };
-
-    let list_model = quote! {
-        #[derive(#list_derives)]
-        pub struct #list_name {
-            #(#list_struct_fields),*
-        }
-
-        impl From<#api_struct_name> for #list_name {
-            fn from(model: #api_struct_name) -> Self {
-                Self {
-                    #(#list_from_assignments),*
-                }
-            }
-        }
-
-        impl From<#struct_name> for #list_name {
-            fn from(model: #struct_name) -> Self {
-                Self {
-                    #(#list_from_model_assignments),*
-                }
-            }
-        }
-    };
-
-    // Generate Response model
-    let response_name = format_ident!("{}Response", api_struct_name);
-    let response_struct_fields =
-        crate::codegen::models::response::generate_response_struct_fields(&raw_fields);
-    let response_from_assignments =
-        crate::codegen::models::response::generate_response_from_assignments(&raw_fields);
-
-    let response_derives =
-        quote! { Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize, utoipa::ToSchema };
-
-    let response_model = quote! {
-        #[derive(#response_derives)]
-        pub struct #response_name {
-            #(#response_struct_fields),*
-        }
-
-        impl From<#api_struct_name> for #response_name {
-            fn from(model: #api_struct_name) -> Self {
-                Self {
-                    #(#response_from_assignments),*
-                }
-            }
-        }
-    };
-
-    (list_model, response_model)
-}
-
 /// # Panics
 ///
 /// This function will panic in the following cases:
@@ -329,7 +259,12 @@ pub fn entity_to_models(input: TokenStream) -> TokenStream {
 
     // Generate list and response models
     let (list_model, response_model) =
-        generate_list_and_response_models(&input, &api_struct_name, struct_name, &field_analysis);
+        codegen::models::list_response::generate_list_and_response_models(
+            &input,
+            &api_struct_name,
+            struct_name,
+            &field_analysis,
+        );
 
     // Generate final output
     let expanded = quote! {
